refactor recent transactions
