+++ pages/api/categories/[userEmail].jsx
