+++ pages/api/transactions/index.js
@@ -4,8 +4,6 @@ export default async function handler(req, res) {
   if (req.method === "POST") {
     const data = req.body;
 
-    console.log(data);
-
     const {
       transactionType,
       transactionName,
